@@ -25,6 +25,8 @@ torchvision = "^0.9.0"
 numba = "^0.53.0"
 scikit-learn = "^0.24.1"
 palkit = "^0.1.7"
+umap-learn = "^0.5.1"
+pretrainedmodels = "^0.7.4"
 
 [tool.poetry.dev-dependencies]
 black = "^20.8b1"
@@ -34,7 +36,7 @@ flake8 = "^3.8.4"
 
 [tool.black]
 line-length = 100
-target-version = ['py38']
+target-version = ['py39']
 skip-string-normalization = true
 include = '\.pyi?$'
 exclude = '''
