@@ -1,6 +1,6 @@
 from .common import *
-from .gmm import *
 from .kmeans import *
 from .pseudo_labelers import *
+from .sklearn import *
 from .tomato import *
 from .topograd import *
