@@ -1,16 +1,15 @@
 from __future__ import annotations
 import logging
 import math
-from typing import Any, Type
+from typing import cast
 
-import matplotlib.pyplot as plt
 import numpy as np
 import torch
 from torch import Tensor
-import torch.nn as nn
 from tqdm import tqdm
 
-from topocluster.clustering.utils import l2_centroidal_distance
+from kit import implements
+from topocluster.clustering.common import Clusterer
 from topocluster.data.datamodules import DataModule
 from topocluster.models.base import Encoder
 from topocluster.utils.torch_ops import compute_density_map, compute_rips
@@ -22,7 +21,9 @@ __all__ = ["topograd_loss", "TopoGrad"]
 LOGGER = logging.getLogger(__name__)
 
 
-def topograd_loss(pc: Tensor, k_kde: int, k_rips: int, scale: float, destnum: int) -> Tensor:
+def topograd_loss(
+    pc: Tensor, k_kde: int, k_rips: int, scale: float, destnum: int
+) -> dict[str, Tensor]:
     kde_dists, _ = compute_density_map(pc, k_kde, scale)
 
     sorted_idxs = torch.argsort(kde_dists, descending=False)
@@ -49,26 +50,27 @@ def topograd_loss(pc: Tensor, k_kde: int, k_rips: int, scale: float, destnum: in
         )
     if not pd_pairs:
         LOGGER.info(
-            "FIltering failed to yield any persistence pairs required for computation of "
+            "Filtering failed to yield sufficient persistence pairs for computation of "
             "the topological loss. Returning 0 instead."
         )
-        return pc.new_zeros(())
-    pd_pairs = torch.as_tensor(pd_pairs, device=pc.device)
-    oripd = kde_dists_sorted[pd_pairs]
-    pers_idxs_sorted = torch.argsort(oripd[:, 0] - oripd[:, 1])
+        shrinking_loss = saliency_loss = pc.new_zeros(())
+    else:
+        pd_pairs = torch.as_tensor(pd_pairs, device=pc.device)
+        oripd = kde_dists_sorted[pd_pairs]
+        pers_idxs_sorted = torch.argsort(oripd[:, 0] - oripd[:, 1])
 
-    changing = pers_idxs_sorted[:-destnum]
-    nochanging = pers_idxs_sorted[-destnum:-1]
+        changing = pers_idxs_sorted[:-destnum]
+        nochanging = pers_idxs_sorted[-destnum:-1]
 
-    biggest = oripd[pers_idxs_sorted[-1]]
-    dest = torch.as_tensor([biggest[0], biggest[1]], device=pc.device)
-    changepairs = pd_pairs[changing]
-    nochangepairs = pd_pairs[nochanging]
-    pd11 = kde_dists_sorted[changepairs]
+        biggest = oripd[pers_idxs_sorted[-1]]
+        dest = torch.as_tensor([biggest[0], biggest[1]], device=pc.device)
+        changepairs = pd_pairs[changing]
+        nochangepairs = pd_pairs[nochanging]
+        pd11 = kde_dists_sorted[changepairs]
 
-    weakdist = torch.sum(pd11[:, 0] - pd11[:, 1]) / math.sqrt(2)
-    strongdist = torch.sum(torch.norm(kde_dists_sorted[nochangepairs] - dest, dim=1))
-    return weakdist + strongdist
+        shrinking_loss = torch.sum(pd11[:, 0] - pd11[:, 1]) / math.sqrt(2)
+        saliency_loss = torch.sum(torch.norm(kde_dists_sorted[nochangepairs] - dest, dim=1))
+    return {"shrinking_loss": shrinking_loss, "saliency_loss": saliency_loss}
 
 
 class TopoGrad(Tomato):
@@ -80,43 +82,47 @@ class TopoGrad(Tomato):
         k_rips: int,
         scale: float,
         threshold: float,
-        iters: int = 0,
+        n_iter: int = 0,
         lr: float = 1e-3,
+        sal_loss_w: float = 1.0,
+        shrink_loss_w: float = 1.0,
     ):
         super().__init__(k_kde=k_kde, k_rips=k_rips, scale=scale, threshold=threshold)
-        self.iters = iters
+        self.n_iter = n_iter
         self.optimizer_cls = torch.optim.AdamW
         self.lr = lr
+        self.sal_loss_w = sal_loss_w
+        self.shrink_loss_w = shrink_loss_w
 
+    @implements(Clusterer)
     def build(self, encoder: Encoder, datamodule: DataModule) -> None:
         self.destnum = datamodule.num_subgroups * datamodule.num_classes
 
+    @implements(Clusterer)
     def _get_loss(self, x: Tensor) -> dict[str, Tensor]:
         if not hasattr(self, "destnum"):
             raise AttributeError(
                 "destnum has not yet been set. Please call 'build' before calling 'get_loss'"
             )
-        loss = topograd_loss(
+        loss_dict = topograd_loss(
             pc=x, k_kde=self.k_kde, k_rips=self.k_rips, scale=self.scale, destnum=self.destnum
         )
-        return {"saliency_loss": loss}
+        loss_dict["saliency_loss"] *= self.sal_loss_w
+        loss_dict["shrinking_loss"] *= self.shrink_loss_w
+        return loss_dict
 
-    def __call__(self, x: Tensor, threshold: float | None = None) -> tuple[Tensor, Tensor]:
+    @implements(Clusterer)
+    def __call__(self, x: Tensor, threshold: float | None = None) -> Tensor:
         threshold = self.threshold if threshold is None else threshold
         # Run topograd on the embedding (without backpropagating through the network)
-        if self.iters > 0:
+        if self.n_iter > 0:
             # Avoid modifying the original embedding
-            x = x.detach().clone().requires_grad_(True)
+            x = x.detach()
+            x = x.clone().requires_grad_(True)
             optimizer = self.optimizer_cls((x,), lr=self.lr)
-            with tqdm(desc="topograd", total=self.iters) as pbar:
-                for _ in range(self.iters):
-                    loss = topograd_loss(
-                        pc=x,
-                        k_kde=self.k_kde,
-                        k_rips=self.k_rips,
-                        scale=self.scale,
-                        destnum=self.destnum,
-                    )
+            with tqdm(desc="topograd", total=self.n_iter) as pbar:
+                for _ in range(self.n_iter):
+                    loss = cast(Tensor, sum(self.get_loss(x=x).values()))
                     optimizer.zero_grad()
                     loss.backward()
                     optimizer.step()
@@ -128,7 +134,7 @@ class TopoGrad(Tomato):
             k_kde=self.k_kde,
             k_rips=self.k_rips,
             scale=self.scale,
-            threshold=self.threshold,
+            threshold=threshold,
         )
         cluster_labels = np.empty(x.shape[0])
         for k, v in enumerate(clusters.values()):
@@ -137,8 +143,5 @@ class TopoGrad(Tomato):
         self.pers_pairs = torch.as_tensor(pers_pairs)
 
         cluster_labels = torch.as_tensor(cluster_labels, dtype=torch.long)
-        centroids = x[list(clusters.keys())]
-        soft_labels = l2_centroidal_distance(x=x, centroids=centroids)
-        hard_labels = cluster_labels
 
-        return hard_labels, soft_labels
+        return cluster_labels
