@@ -1,17 +1,19 @@
 from __future__ import annotations
 from abc import abstractmethod, abstractstaticmethod
-from typing import ClassVar, cast
+from typing import Any, ClassVar, List, cast
 
 import pytorch_lightning as pl
 import torch
 from torch.tensor import Tensor
 from torch.utils.data import DataLoader
 from torch.utils.data.dataset import Dataset, Subset
+from torch.utils.data.sampler import RandomSampler, Sampler
 from torchvision import transforms
 from torchvision.datasets import MNIST
 
 from kit import implements
 from kit.torch import prop_random_split
+from topocluster.data.sampling import GreedyCoreSetSampler
 from topocluster.data.utils import (
     Batch,
     BinarizedLabelDataset,
@@ -43,6 +45,7 @@ class DataModule(pl.LightningDataModule):
         test_batch_size: int = 1000,
         val_batch_size: int | None = None,
         num_workers: int = 0,
+        train_batch_sampler: Sampler[List[int]] | None = None,
     ):
         super().__init__()
         self.data_dir = data_dir
@@ -51,6 +54,7 @@ class DataModule(pl.LightningDataModule):
         self.val_batch_size = test_batch_size if val_batch_size is None else val_batch_size
         self.num_workers = num_workers
         self._collate_fn = cast_collation(adaptive_collate, Batch)
+        self.train_batch_sampler = train_batch_sampler
 
     @property
     @abstractmethod
@@ -63,15 +67,20 @@ class DataModule(pl.LightningDataModule):
         ...
 
     @implements(pl.LightningDataModule)
-    def train_dataloader(self) -> DataLoader:
+    def train_dataloader(self, shuffle: bool = True) -> DataLoader:
+        dl_kwargs: dict[str, Any] = {
+            "shuffle": shuffle and self.train_batch_sampler is None,
+            "drop_last":self.train_batch_sampler is None
+        }
+        if self.train_batch_sampler is None:
+            dl_kwargs["batch_size"] = self.train_batch_size
         return DataLoader(
             self.train_data,
-            batch_size=self.train_batch_size,
-            shuffle=True,
             pin_memory=True,
             num_workers=self.num_workers,
-            drop_last=True,
             collate_fn=self._collate_fn,
+            batch_sampler=self.train_batch_sampler,
+            **dl_kwargs
         )
 
     @implements(pl.LightningDataModule)
