@@ -5,6 +5,7 @@ from torch import Tensor
 import torch
 import torch.nn as nn
 
+from kit import implements
 from topocluster.data.datamodules import VisionDataModule
 from topocluster.data.utils import Batch
 from topocluster.models.base import Encoder
@@ -17,7 +18,8 @@ class LeNet4(Encoder):
     Adapted from https://github.com/activatedgeek/LeNet-5
     """
 
-    def _build(self, datamodule: VisionDataModule) -> nn.Module:
+    @implements(Encoder)
+    def _build(self, datamodule: VisionDataModule) -> tuple[nn.Module, int]:
 
         encoder = nn.Sequential(
             nn.LazyConv2d(6, kernel_size=(5, 5)),
@@ -34,9 +36,10 @@ class LeNet4(Encoder):
         self.fc = nn.LazyLinear(logits_dim)
         self.criterion = nn.CrossEntropyLoss() if logits_dim > 2 else nn.BCEWithLogitsLoss()
         # Lazy initialization
-        self.fc(encoder(torch.ones(datamodule.dims)[None]))
-        return encoder
+        self.fc(latent_dim := encoder(torch.ones(datamodule.dims)[None]))
+        return encoder, latent_dim
 
+    @implements(Encoder)
     def _get_loss(self, encoding: Tensor, batch: Batch) -> dict[str, Tensor]:
         logits = self.fc(encoding)
         targets = batch.y
